Move alterodo find strategy
